clean main file, add functions file
